+++ cursor.go
@@ -4,25 +4,16 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"log"
-	"strconv"
 	"strings"
-)
 
-type cursor struct {
-	ProjectID string `json:"projectId"`
-	Signature string `json:"signature"`
-	Cursor    string `json:"cursor"`
-}
+	"cloud.google.com/go/datastore"
+)
 
 // Cursor :
 type Cursor struct {
-	cc []byte
-}
-
-func (c Cursor) offset() int32 {
-	i, _ := strconv.Atoi(strings.Replace(string(c.cc), "offset=", "", -1))
-	return int32(i)
+	cc        []byte
+	Signature string         `json:"signature"`
+	Key       *datastore.Key `json:"next"`
 }
 
 // String :
@@ -45,8 +36,10 @@ func DecodeCursor(c string) (Cursor, error) {
 	if err != nil {
 		return Cursor{}, fmt.Errorf("goloquent: invalid cursor")
 	}
-	var m map[string]interface{}
-	json.Unmarshal(b, &m)
-	log.Println(m)
-	return Cursor{b}, nil
+	cc := new(Cursor)
+	cc.cc = b
+	if err := json.Unmarshal(b, cc); err != nil {
+		return Cursor{}, fmt.Errorf("goloquent: invalid cursor")
+	}
+	return *cc, nil
 }