+++ relationship.go
@@ -1,9 +1,0 @@
-package goloquent
-
-// Relationship :
-type Relationship struct {
-	ForeignKey string
-	Parent     string
-	Relation   string
-	Child      string
-}