+++ statement.go
@@ -3,8 +3,8 @@ package goloquent
 import (
 	"bytes"
 	"database/sql"
-	"fmt"
 	"strings"
+	"time"
 )
 
 type stmt struct {
@@ -12,11 +12,11 @@ type stmt struct {
 	arguments []interface{}
 }
 
-func (s *stmt) string() string {
+func (s stmt) string() string {
 	return s.statement.String()
 }
 
-func (s *stmt) canSkip() bool {
+func (s stmt) isZero() bool {
 	return !(s.statement.Len() > 0)
 }
 
@@ -28,18 +28,30 @@ type replacer interface {
 // Stmt :
 type Stmt struct {
 	stmt
-	Result   sql.Result
-	replacer replacer
+	crud      string
+	replacer  replacer
+	startTime time.Time
+	endTime   time.Time
+	Result    sql.Result
 }
 
 func (s *Stmt) startTrace() {
+	s.startTime = time.Now().UTC()
+}
 
+func (s *Stmt) stopTrace() {
+	s.endTime = time.Now().UTC()
+}
+
+// TimeElapse :
+func (s Stmt) TimeElapse() time.Duration {
+	return s.endTime.Sub(s.startTime)
 }
 
 // Raw :
 func (s *Stmt) Raw() string {
 	buf := new(bytes.Buffer)
-	arr := strings.Split(s.string(), "??")
+	arr := strings.Split(s.string(), variable)
 	for i := 0; i < len(arr); i++ {
 		str := arr[i] + s.replacer.Bind(uint(i+1))
 		if i >= len(arr)-1 {
@@ -53,7 +65,7 @@ func (s *Stmt) Raw() string {
 // String :
 func (s *Stmt) String() string {
 	buf := new(bytes.Buffer)
-	arr := strings.Split(s.string(), "??")
+	arr := strings.Split(s.string(), variable)
 	for i, aa := range s.arguments {
 		str := arr[i] + s.replacer.Value(aa)
 		buf.WriteString(str)
@@ -63,19 +75,6 @@ func (s *Stmt) String() string {
 }
 
 // Arguments :
-func (s *Stmt) Arguments() []interface{} {
+func (s Stmt) Arguments() []interface{} {
 	return s.arguments
-}
-
-func toString(it interface{}) string {
-	var str string
-	switch vi := it.(type) {
-	case nil:
-		str = "NULL"
-	case string, []byte:
-		str = fmt.Sprintf("%q", vi)
-	default:
-		str = fmt.Sprintf("%v", vi)
-	}
-	return str
 }